[test] merge TestRaiseException cases and run on main suite
