+++ core/src/test/java/org/jruby/test/ThrowFromJava.java
@@ -1,0 +1,5 @@
+package org.jruby.test;
+
+public class ThrowFromJava {
+    public void throwIt() { throw new RuntimeException("here"); }
+}
